Updated scripts for quickstart
The quickstart scripts were assuming a root.war file existed.
The 50-jetty.bash script was intercepting too many command lines, including quickstart ones.
The 60-jetty-dry-run.bash script was ignoring when arguments changed, renamed to 60-jetty-start.bash
The generate-jetty-start.sh script now uses the 60-jetty-start.bash script to do the actual generation and the gcp module is enabled for that.
Tests needed to be updated to match